==> 4.1.Matching_triplex.ipynb <==
# Matching

### Preliminary Operations

import pandas as pd
import numpy as np

#pd.set_option('display.max_rows', None)

df_triple_kb = pd.read_csv(r'df_triple_KB.csv')
df_triple_kb

df_triple_test = pd.read_csv(r'df_triple_TEST.csv')
df_triple_test

## Matching Triple
