==> 4.2 Matching_Embedding.ipynb <==
# Semantic Matching  

## Preliminary Operations

import pandas as pd
import numpy as np
import faiss
import openai
from openai import OpenAI
from sklearn.metrics import f1_score, accuracy_score, precision_score, confusion_matrix
import re

#pd.set_option('display.width', 1000) 
#pd.set_option('display.max_colwidth', None)
#pd.set_option('display.max_rows', None)

client = OpenAI(api_key="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")

df_kb = pd.read_csv(r'df_kb_generation.csv')
df_test = pd.read_csv(r'df_test_generation.csv', nrows=20)

unique_label_counts = df_test['label'].value_counts()
print(unique_label_counts)

unique_label_counts = df_kb['label'].value_counts()
print(unique_label_counts)

#Clean text
# Rimuovi le parti precedenti a :\r\n
df_test['new_claim'] = df_test['new_claim'].apply(lambda x: re.sub(r'.*:\r\n', '', x)) 
df_test['new_expl'] = df_test['new_expl'].apply(lambda x: re.sub(r'.*:\r\n', '', x))

#Clean text
#Rimuovi altri parti
df_test['new_claim'] = df_test['new_claim'].replace('"""\r\n', '', regex=True) 
df_test['new_claim'] = df_test['new_claim'].replace('\r\n"""', '', regex=True)
df_test['new_expl'] = df_test['new_expl'].replace('"""\r\n', '', regex=True) 
df_test['new_expl'] = df_test['new_expl'].replace('\r\n"""', '', regex=True)

#Clean text
# Rimuovi le parti precedenti a :\r\n
df_kb['new_claim'] = df_kb['new_claim'].apply(lambda x: re.sub(r'.*:\r\n', '', x)) 
df_kb['new_expl'] = df_kb['new_expl'].apply(lambda x: re.sub(r'.*:\r\n', '', x))

#Clean text
#Rimuovi altri parti
df_kb['new_claim'] = df_kb['new_claim'].replace('"""\r\n', '', regex=True) 
df_kb['new_claim'] = df_kb['new_claim'].replace('\r\n"""', '', regex=True)
df_kb['new_expl'] = df_kb['new_expl'].replace('"""\r\n', '', regex=True) 
df_kb['new_expl'] = df_kb['new_expl'].replace('\r\n"""', '', regex=True)

df_kb[['new_claim','new_expl']]

df_test[['new_claim','new_expl']]

### Matching New Claim

claims = df_kb["new_claim"].tolist()
len(claims)

def get_embedding(input, model="text-embedding-3-small", encoding_format = "float"):
  obj = client.embeddings.create(
    model=model,
    input=input,
    encoding_format=encoding_format
  )
  return obj.data[0].embedding

import pickle
with open("embeddings_claims.pkl", "rb") as f:
    vectors_claim = pickle.load(f)

vectors_claim.shape[1]

np.arange(len(claims))

index = faiss.IndexIDMap(faiss.IndexFlatIP(vectors_claim.shape[1]))
faiss.normalize_L2(vectors_claim)
index.add_with_ids(vectors_claim, np.arange(len(claims)))

def index_search(query, k=1):
    query_vector =get_embedding(query)
    query_vector = np.array(query_vector, dtype=np.float32).reshape(1, -1)
    faiss.normalize_L2(query_vector)
    top_k = index.search(query_vector, k)
    
    return(top_k[0], [
        claims[_id] for _id in top_k[1][0]
    ])

index_search('Cigarette cause cancer', k=3)

def verify_claim(claim, k=3, thresholds=[0.5, 0.7, 0.9]):
    # Ottieni i risultati della ricerca FAISS per la nuova affermazione
    similarity_scores, matched_claims = index_search(claim, k=k)
    
    # La similarità massima trovata tra la nuova affermazione e le corrispondenze
    max_similarity = np.max(similarity_scores) if similarity_scores.size > 0 else 0
    
    # Determina se l'affermazione è vera secondo le soglie di similarità
    verdicts = {}
    for threshold in thresholds:
        verdicts[threshold] = "True" if max_similarity >= threshold else "False"
    
    return verdicts, max_similarity, matched_claims

# Esempio di utilizzo della funzione
claim = 'Cigarette cause cancer'
verdicts, max_similarity, matched_claims = verify_claim(claim, k=3)

print(f"Verdicts: {verdicts}")
print(f"Max Similarity: {max_similarity}")
print("Matched Claims:")
for claim in matched_claims:
    print(f"- {claim}")

def get_binary_verdict_and_print(row, threshold=0.5):
    claim = row['new_claim']
    true_label = row['label']  # Assume che la colonna con le label si chiami 'label'
    _, max_similarity, matched_claims = verify_claim(claim, k=3)
    is_true = max_similarity >= threshold

    print(f"Claim: {claim}")
    print(f"True Label: {true_label}")
    print(f"Verdict at threshold {threshold}: {'True' if is_true else 'False'}")
    print("Matched Claims from Knowledge Base:")
    for matched_claim in matched_claims:
        print(f"- {matched_claim}")
    print(f"Max Similarity: {max_similarity}\n")
    
    return is_true

# Applica la funzione apply_verify_and_get_binary_verdict a ogni new_claim nel DataFrame
df_test['predicted_label'] = df_test.apply(get_binary_verdict_and_print, threshold=0.5, axis=1)

true_labels = df_test['label'].apply(lambda x: 'true' if x == 'true' else 'false')
true_labels = true_labels.map({'true': 1, 'false': 0})
#predicted_labels = df_test['predicted_label'].fillna('false') # Sostituisce NaN con 'false'
predicted_labels = df_test['predicted_label'].map({True: 1, False: 0})  

# Calcolo delle metriche
f1 = f1_score(true_labels, predicted_labels)
accuracy = accuracy_score(true_labels, predicted_labels)
precision = precision_score(true_labels, predicted_labels, zero_division=0)

print(f"F1-Score: {f1}")
print(f"Accuracy: {accuracy}")
print(f"Precision: {precision}")

cm_a = confusion_matrix(true_labels, predicted_labels)
    
tn_a, fp_a, fn_a, tp_a = cm_a.ravel()

print(f"\nMetodo 'a' - Matrice di Confusione:")
print(f"TN: {tn_a}, FP: {fp_a}, FN: {fn_a}, TP: {tp_a}")

### Matching New Explanation

explanation = df_kb["new_expl"].tolist()
len(explanation)

def get_embedding(input, model="text-embedding-3-small", encoding_format = "float"):
  obj = client.embeddings.create(
    model=model,
    input=input,
    encoding_format=encoding_format
  )
  return obj.data[0].embedding

with open("embeddings_explaination.pkl", "rb") as f:
    vectors_expl = pickle.load(f)

vectors_expl.shape[1]

np.arange(len(claims))

index = faiss.IndexIDMap(faiss.IndexFlatIP(vectors_expl.shape[1]))
faiss.normalize_L2(vectors_expl)
index.add_with_ids(vectors_expl, np.arange(len(explanation)))

def index_search(query, k=1):
    query_vector =get_embedding(query)
    query_vector = np.array(query_vector, dtype=np.float32).reshape(1, -1)
    faiss.normalize_L2(query_vector)
    top_k = index.search(query_vector, k)
    
    return(top_k[0], [
        explanation[_id] for _id in top_k[1][0]
    ])

def verify_new_expl(claim, k=3, thresholds=[0.5, 0.7, 0.9]):
    # Ottieni i risultati della ricerca FAISS per la nuova affermazione
    similarity_scores, matched_claims = index_search(claim, k=k)
    
    # La similarità massima trovata tra la nuova affermazione e le corrispondenze
    max_similarity = np.max(similarity_scores) if similarity_scores.size > 0 else 0
    
    # Determina se l'affermazione è vera secondo le soglie di similarità
    verdicts = {}
    for threshold in thresholds:
        verdicts[threshold] = "True" if max_similarity >= threshold else "False"
    
    return verdicts, max_similarity, matched_claims

index_search('Tobacco cause cancer', k=3)

def get_binary_verdict_and_print(row, threshold=0.5):
    new_expl = row['new_expl']
    true_label = row['label']  # Assume che la colonna con le label si chiami 'label'
    _, max_similarity, matched_new_expls = verify_new_expl(new_expl, k=3)
    is_true = max_similarity >= threshold

    print(f"Explanation: {new_expl}")
    print(f"True Label: {true_label}")
    print(f"Verdict at threshold {threshold}: {'True' if is_true else 'False'}")
    print("Matched new_expls from Knowledge Base:")
    for matched_new_expl in matched_new_expls:
        print(f"- {matched_new_expl}")
    print(f"Max Similarity: {max_similarity}\n")
    
    return is_true

# Applica la funzione apply_verify_and_get_binary_verdict a ogni new_claim nel DataFrame
df_test['predicted_label'] = df_test.apply(get_binary_verdict_and_print, threshold=0.5, axis=1)

true_labels = df_test['label'].apply(lambda x: 'true' if x == 'true' else 'false')
true_labels = true_labels.map({'true': 1, 'false': 0})
#predicted_labels = df_test['predicted_label'].fillna('false') # Sostituisce NaN con 'false'
predicted_labels = df_test['predicted_label'].map({True: 1, False: 0})  

# Calcolo delle metriche
f1 = f1_score(true_labels, predicted_labels)
accuracy = accuracy_score(true_labels, predicted_labels)
precision = precision_score(true_labels, predicted_labels, zero_division=0)

print(f"F1-Score: {f1}")
print(f"Accuracy: {accuracy}")
print(f"Precision: {precision}")

cm_a = confusion_matrix(true_labels, predicted_labels)
    
tn_a, fp_a, fn_a, tp_a = cm_a.ravel()

print(f"\nMetodo 'a' - Matrice di Confusione:")
print(f"TN: {tn_a}, FP: {fp_a}, FN: {fn_a}, TP: {tp_a}")
